Add model utils
